requirements.txt
Django==1.5.1
South==0.8.1
argparse==1.2.1
distribute==0.6.34
dj-database-url==0.2.1
django-braces==1.0.0
django-crispy-forms==1.3.2
django-discover-runner==1.0
django-floppyforms==1.1
django-model-utils==1.4.0
psycopg2==2.5.1
wsgiref==0.1.2
